+++ BankService/Deal/src/main/java/com/munsun/deal/utils/filters/LoggingRequestResponseFilter.java
@@ -1,0 +1,45 @@
+package com.munsun.deal.utils.filters;
+
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import org.springframework.web.util.ContentCachingRequestWrapper;
+import org.springframework.web.util.ContentCachingResponseWrapper;
+
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+
+@Slf4j
+@Component
+public class LoggingRequestResponseFilter implements Filter {
+    @Override
+    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+        var request = (HttpServletRequest) servletRequest;
+        var response = (HttpServletResponse) servletResponse;
+
+        ContentCachingRequestWrapper requestWrapper = new ContentCachingRequestWrapper(request);
+        ContentCachingResponseWrapper responseWrapper = new ContentCachingResponseWrapper(response);
+
+        filterChain.doFilter(requestWrapper, responseWrapper);
+
+        String requestBody = getStringValue(requestWrapper.getContentAsByteArray(),
+                request.getCharacterEncoding());
+        String responseBody = getStringValue(responseWrapper.getContentAsByteArray(),
+                response.getCharacterEncoding());
+
+        log.info("Request{{} {}, body={}}; Response{{}; body={}}",
+                request.getMethod(), request.getRequestURI(), requestBody, response.getStatus(), responseBody);
+        responseWrapper.copyBodyToResponse();
+    }
+
+    private String getStringValue(byte[] contentAsByteArray, String characterEncoding) {
+        try {
+            return new String(contentAsByteArray, 0, contentAsByteArray.length, characterEncoding);
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
+        }
+        return "";
+    }
+}