+++ BankService/Deal/changelog.md
@@ -1,12 +1,21 @@
+# 1.1.3 - 2025.03.04
+
+### Patch:
+
+- удалены перехватчики для логирования запросов/ответов;
+- добавлены аспекты с счетчиками для подсчета заявок со статусом APPROVED и CREDIT_ISSUED. [Статья](https://habr.com/ru/companies/otus/articles/650871/), [Статья](https://docs.micrometer.io/micrometer/reference/concepts/counters.html);
+- добавлен ErrorDecoder для feign client.
+
+
 # 1.1.2 - 2025.03.03
 
-### Fix: 
+### Patch: 
 
 - поменял название параметра в feign client
 
 # 1.1.1 - 2025.03.03
 
-### Fix:
+### Patch:
 
 - поменял свойство для клиента
 
@@ -18,7 +27,7 @@
 
 # 1.0.0 - 2025.03.02
 
-### Refactor:
+### Patch:
 
 - добавлена генерация моделей с помощью плагина open-api-maven-plugin;